Fix bugs

+++ src/monitoring/create_tables/batch_timestamps.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS batch_timestamps (
     stage String NOT NULL,
     status String NOT NULL,
     timestamp DateTime64(6) NOT NULL,
-    message_count UInt64
+    message_count UInt32
 )
 ENGINE = MergeTree
 PRIMARY KEY (batch_id);
